refactor(tests): consolidate domain tests from 14 to 12 files
Merge overlapping test files to reduce duplication:
- Merge events/test_events.py into test_events.py (8 tests)
- Merge handlers/test_manager.py into handlers/test_issue_state_dispatch.py (11 tests)
- Remove duplicate tests from test_qualify_gate.py (covered by separate files)

Deleted files:
- tests/vibe3/domain/events/test_events.py (duplicate tests)
- tests/vibe3/domain/events/ directory (now empty)
- tests/vibe3/domain/handlers/test_manager.py (overlap with test_issue_state_dispatch.py)

Preserved files (with updated BlockedStateService integration from main):
- tests/vibe3/domain/test_qualify_gate_remote.py (remote coordination tests)
- tests/vibe3/domain/test_qualify_gate_service_calls.py (service call tests)

Result: 14 → 12 test files (-14% reduction)
All 80 tests pass, lint clean, no regression.
